+++ packages/medical-domain/test/repositories/statistics/insulin-statistics.service.test.ts
@@ -94,16 +94,16 @@ describe('getBasalBolusData', () => {
 describe('getManualBolusAverageStatistics', () => {
   it('should return a map with ranges of hours and manual bolus average statistics', () => {
     const boluses = buildBolusData(manualBolusData)
-    const manualBoluses = BasalBolusStatisticsService.getManualBolusAverageStatistics(boluses, 14, dateFilterTwoWeeks)
+    const manualBoluses = BasalBolusStatisticsService.getManualBolusAverageStatistics(boluses, dateFilterTwoWeeks)
     const expected: ManualBolusAverageStatistics = new Map([
-      [HoursRange.MidnightToThree, { confirmedDose: 0.1, numberOfInjections: 0.1 }],
-      [HoursRange.ThreeToSix, { confirmedDose: 0.4, numberOfInjections: 0.1 }],
+      [HoursRange.MidnightToThree, { confirmedDose: 3, numberOfInjections: 2 }],
+      [HoursRange.ThreeToSix, { confirmedDose: 5, numberOfInjections: 1 }],
       [HoursRange.SixToNine, { confirmedDose: 0, numberOfInjections: 0 }],
       [HoursRange.NineToTwelve, { confirmedDose: 0, numberOfInjections: 0 }],
-      [HoursRange.TwelveToFifteen, { confirmedDose: 0.3, numberOfInjections: 0.1 }],
+      [HoursRange.TwelveToFifteen, { confirmedDose: 2, numberOfInjections: 2 }],
       [HoursRange.FifteenToEighteen, { confirmedDose: 0, numberOfInjections: 0 }],
-      [HoursRange.EighteenToTwentyOne, { confirmedDose: 0.5, numberOfInjections: 0.1 }],
-      [HoursRange.TwentyOneToMidnight, { confirmedDose: 0.3, numberOfInjections: 0.2 }]
+      [HoursRange.EighteenToTwentyOne, { confirmedDose: 4, numberOfInjections: 2 }],
+      [HoursRange.TwentyOneToMidnight, { confirmedDose: 1, numberOfInjections: 3 }]
     ])
     expect(manualBoluses).toEqual(expected)
   })